Review changes for opensearch restore
